+++ HISTORY.rst
@@ -2,6 +2,18 @@
 History
 =======
 
+0.4.0 (2017-03-28)
+------------------
+
+* Add picard output data
+* Add functions for dealing with application fixtures
+* Move docs to gh-pages
+* Add pytest_ngsfixtures_add_application.py for templating new
+  applications
+* Application outputs now implemented as dictionaries for
+  multiple-output applications
+
+
 0.3.1 (2017-03-03)
 ------------------
 
@@ -20,7 +32,7 @@ Add application data, pool data, and functions for downloading data
 * Add application data for cutadapt, fastqc, qualimap and samtools
 * Add pooled sequencing data
 * Add functionality for downloading large data sets
-* Improve make targets for easier releases  
+* Improve make targets for easier releases
 * Minor bug fixes
 
 
@@ -36,7 +48,7 @@ Add new reference files, test files, and factory functions.
 * Add threads options for running tests
 * Add Snakefile and rules for generating application files
 * Add fileset, filetype factory functions
-  
+
 
 0.1.0 (2017-01-24)
 ------------------